LE-40 Upon clicking marker, bring up appropriate editor for the location
* Adds set of tabbed pages for each of three Location states.
* Extends the Leaflet marker with a Location ID that can be used to
retrieve the location when a click event fires.
* Stores Locations in a map for retrieval.
* Adds logic to figure out which tab is appropriate for a given location's
readiness level.

The connection between the event-response and the Nav Controller is tricky
because 'this' is defined for the CRMarker rather than the class defining
the response.  An anonymous function is being used, but may want to explore
using the retrieval of the Root Nav instead.