LE-82 Markers have problems
This is not the solution to the problem, but while testing out this
feature, I ran into the problem where "ionic serve" would abort when
the page was refreshed.  The solution was to rollback the version of
the 'ws' module to 3.3.2.

The actual solution was to follow the README.md in the markers package:
"Copy leaflet.awesome-marker's CSS to a SCSS file."