LE-95 App State and page for tracking/displaying the status of preparing the application
- Refactors much of the sequencing into a new AppStateService.
- Changes dependency hierarchy to avoid circular dependencies.
- Adds PlatformService to tell if we're running native or not.
- Adds Status page to display summary of the App State.
- Opens a new Map Data service for handling the state of Map Data;
shell at this time.

This takes the next step toward increased usage of RxJS. One notable
addition is turning the Platform Ready response into a shared observable
that also logs its occurrence.

Part of this pattern is the use of the '=>' arrow function to localize
the 'this' reference.

Also removes the unused session-token resource.