LE-97 Assure registration is complete before kicking off other activity
Although this code wasn't technically the root cause, it does show
that I had a gap in my understanding of how these pieces tied together.

The changes here are improvements and movement toward a declarative
approach over an imperative approach. Reach into the stream and take
actions as the marbles come by.