LE-102 Tie "Use this Email" to the status page's flow
- Sets up paying attention to the Confirmation State to know when we
can proceed after setting up registration.

Response includes popping out of Confirmation and Registration pages.

Now that I see this, we may not need to have the separate apps know
about this. They could setup the page that we want to serve as the "Home"
page and then let the FEC get us back to that page.